requirements.txt
Django~=3.1.1
djangorestframework~=3.12.1
django-cors-headers==3.5.0
psycopg2-binary==2.8.6